(refactor) : 암호화 패키지, 인터페이스 변경 #EAT-38
