--- src/main/java/team/gwon/haveameal/member/memberinfoencryption/AesMemberInfoEncryptionService.java
@@ -1,7 +1,0 @@
-package team.gwon.haveameal.member.memberinfoencryption;
-
-public interface AesMemberInfoEncryptionService {
-	String encryptInfo(String plainInfo);
-
-	String decryptInfo(String encryptionData);
-}
